requirements.txt
json5
salesforce-lavis==1.0.2
tqdm==4.65.0
